# flight-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple Go service for flight booking management",
        "title": "Flight Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@flight.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of all orders (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.PaginatedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    }
                }
            }
        },
        "/admin/orders/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing order (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Update order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Order data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.Order"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of orders for the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List my orders",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.PaginatedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new flight booking order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.Order"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single order by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.Order"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel an existing order",
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessMessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorMessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.Airline": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "GA"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-12-07T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "ga"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "logo": {
                    "type": "string",
                    "example": "https://example.com/logo.png"
                },
                "name": {
                    "type": "string",
                    "example": "Garuda Indonesia"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-12-07T00:00:00Z"
                }
            }
        },
        "handlers.Airport": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Jakarta"
                },
                "code": {
                    "type": "string",
                    "example": "CGK"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-12-07T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "cgk"
                },
                "name": {
                    "type": "string",
                    "example": "Soekarno-Hatta International Airport"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-12-07T00:00:00Z"
                }
            }
        },
...
        "handlers.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "cabin_class": {
                    "type": "string",
                    "example": "economy"
                },
                "contact_email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "contact_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "contact_phone": {
                    "type": "string",
                    "example": "+6281234567890"
                },
                "flight_date": {
                    "type": "string",
                    "example": "2024-12-20"
                },
                "passengers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.PassengerRequest"
                    }
                },
                "schedule_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "handlers.ErrorMessageResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.Order": {
            "type": "object",
            "properties": {
                "cabin_class": {
                    "type": "string",
                    "example": "economy"
                },
                "contact_email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "contact_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "contact_phone": {
                    "type": "string",
                    "example": "+6281234567890"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-12-07T00:00:00Z"
                },
                "flight_date": {
                    "type": "string",
                    "example": "2024-12-20"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "passengers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Passenger"
                    }
                },
                "schedule": {
                    "$ref": "#/definitions/handlers.Schedule"
                },
                "schedule_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "total_amount": {
                    "type": "number",
                    "example": 3000000
                },
                "total_passenger": {
                    "type": "integer",
                    "example": 2
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-12-07T00:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "handlers.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total_items": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "handlers.Passenger": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "order_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "title": {
                    "type": "string",
                    "example": "Mr"
                },
                "type": {
                    "type": "string",
                    "example": "adult"
                }
            }
        },
        "handlers.PassengerRequest": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "title": {
                    "type": "string",
                    "example": "Mr"
                },
                "type": {
                    "type": "string",
                    "example": "adult"
                }
            }
        },
...
        "handlers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "handlers.Schedule": {
            "type": "object",
            "properties": {
                "aircraft": {
                    "type": "string",
                    "example": "Boeing 737-800"
                },
                "airline": {
                    "$ref": "#/definitions/handlers.Airline"
                },
                "airline_id": {
                    "type": "string",
                    "example": "ga"
                },
                "arrival_airport": {
                    "$ref": "#/definitions/handlers.Airport"
                },
                "arrival_airport_id": {
                    "type": "string",
                    "example": "dps"
                },
                "arrival_terminal": {
                    "type": "string",
                    "example": "D"
                },
                "arrival_time": {
                    "type": "string",
                    "example": "10:30"
                },
                "business_price": {
                    "type": "number",
                    "example": 4500000
                },
                "business_seats": {
                    "type": "integer",
                    "example": 30
                },
                "days_of_week": {
                    "type": "string",
                    "example": "1,2,3,4,5,6,7"
                },
                "departure_airport": {
                    "$ref": "#/definitions/handlers.Airport"
                },
                "departure_airport_id": {
                    "type": "string",
                    "example": "cgk"
                },
                "departure_terminal": {
                    "type": "string",
                    "example": "3"
                },
                "departure_time": {
                    "type": "string",
                    "example": "08:00"
                },
                "duration": {
                    "type": "integer",
                    "example": 150
                },
                "economy_price": {
                    "type": "number",
                    "example": 1500000
                },
                "economy_seats": {
                    "type": "integer",
                    "example": 150
                },
                "first_class_price": {
                    "type": "number",
                    "example": 8000000
                },
                "first_class_seats": {
                    "type": "integer",
                    "example": 10
                },
                "flight_number": {
                    "type": "string",
                    "example": "GA123"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "valid_from": {
                    "type": "string",
                    "example": "2024-01-01"
                },
                "valid_until": {
                    "type": "string",
                    "example": "2024-12-31"
                }
            }
        },
        "handlers.SuccessMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "services.UpdateOrderRequest": {
            "type": "object",
            "properties": {
                "contact_email": {
                    "type": "string"
                },
                "contact_name": {
                    "type": "string"
                },
                "contact_phone": {
                    "type": "string"
                },
                "status": {
                    "description": "pending, confirmed, cancelled, completed",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}